pyproject.toml
[project]
name = "binsmith"
version = "0.1.0"
description = "A bash-native AI agent that builds its own toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
keywords = ["ai", "agent", "cli", "automation", "llm", "bash", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "logfire>=3.0.0",
    "pydantic-ai>=1.37.0",
    "pydantic-ai-slim[ag-ui]>=1.37.0",
    "textual>=6.11.0",
    "trafilatura>=2.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
binsmith = "binsmith.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
